+++ scripts/create-icons-data.script.ts
@@ -2,15 +2,24 @@
  * Script will create json data, which contains dictionaries for runtime,
  * where `key` is name of folder we want icon for and `value` is icon's filename.
  */
-import { script } from './utils';
-import { writeFileSync, readFileSync } from 'fs';
-import Ch from 'chalk';
+import * as log4js from "log4js";
+import { readFileSync, createWriteStream } from 'fs';
+import { basename, extname } from "path"
+
+const log = log4js.getLogger(basename(__filename, extname(__filename)));
+log.level = "debug";
+
+// Open icon definitions
+log.debug("opening icon definitions");
+
 const iconsJSONFile = readFileSync('./data/generated/icons.json');
 const vsiLanguagesFile = readFileSync('./data/generated/languages-vsi.json');
 const vscodeLanguagesFile = readFileSync('./data/static/languages-vscode.json');
 const PATH_ICONSDATA = './src/generated/';
 type IconKey = string;
 
+// Parse icon definitions
+log.debug("parsing icon definitions");
 const icons = JSON.parse(iconsJSONFile.toString()) as {
     iconDefinitions: { [iconKey: string]: { iconPath: string } },
     folderNames: { [folderName: string]: IconKey },
@@ -25,135 +34,133 @@ const icons = JSON.parse(iconsJSONFile.toString()) as {
     }
 };
 
-const languages = JSON.parse(vsiLanguagesFile.toString()) as {
+const vsiLanguages = JSON.parse(vsiLanguagesFile.toString()) as {
     [language: string]: {
         ids: string | string[];
         defaultExtension: string;
     }
 };
 
-const writeFile = (fileName: string, callback: () => any) => {
-    const result = callback();
-    writeFileSync(
-        PATH_ICONSDATA + fileName + '.ts',
-        `export const ${fileName}: { [key: string]: string } = ${JSON.stringify(result, null, 2)}`
-    );
-    console.log(Ch.green(`> '${PATH_ICONSDATA + fileName + '.ts'}' file created`));
+const vscodeLanguages = JSON.parse(vscodeLanguagesFile.toString()) as {
+    [languageId: string]: {
+        extensions: string[],
+        filenames?: string[]
+    }
+};
+
+// Icon To Path
+const iconToPath: { [icon: string]: string } = Object.keys(icons.iconDefinitions).reduce((acc, icon) => ({
+    ...acc,
+    [icon]: icons.iconDefinitions[icon].iconPath.split('/').pop()
+}), {});
+
+
+function getIconPath(icon: string) {
+    return iconToPath[icon];
 }
 
-// create mini-json files
-
-script(__filename, 'Creating mini-json files from definitions', ({ log }, exit) => {
-    const iconToPath: any = Object.keys(icons.iconDefinitions).reduce((acc, icon) => ({
-        ...acc,
-        [icon]: icons.iconDefinitions[icon].iconPath.split('/').pop()
-    }), {});
-
-    // FolderNames to Icon
-    writeFile(`FolderNamesToIcon`, () => {
-        const folderNames = Object.keys(icons.light.folderNames).reduce((acc, folderName) => ({
-            ...acc,
-            [folderName]: iconToPath[
-                icons.light.folderNames[folderName]
-            ]
-        }), {});
-        return folderNames;
-    });
 
 
-    // FileExtensions to Icon
-    writeFile(`FileExtensions1ToIcon`, () => {
-        // 1 - .js, .ts, .cpp
-        const fileExtensions1 = Object.keys(icons.light.fileExtensions).reduce((acc, fileExtension) => {
-            if (fileExtension.indexOf('.') === -1) {
-                return {
-                    ...acc,
-                    [fileExtension]: iconToPath[
-                        icons.light.fileExtensions[fileExtension]
-                    ]
-                };
-            }
-            return { ...acc };
-        }, {});
-        return fileExtensions1;
-    });
-    writeFile(`FileExtensions2ToIcon`, () => {
-        // 2. - .js.map, .test.js, .test.ts
-        const fileExtensions2 = Object.keys(icons.light.fileExtensions).reduce((acc, fileExtension) => {
-            if (fileExtension.indexOf('.') > -1) {
-                return {
-                    ...acc,
-                    [fileExtension]: iconToPath[
-                        icons.light.fileExtensions[fileExtension]
-                    ]
-                };
-            }
-            return { ...acc };
-        }, {});
-        return fileExtensions2;
-    });
+// FolderNames to Icon
+(async function() {
+    log.debug("creating table for foldernames");
+    const folderIcons = createWriteStream(PATH_ICONSDATA + 'FolderNamesToIcon.ts', { flags: 'w'});
+    folderIcons.write(`export const FolderNamesToIcon: { [key: string]: string } = {\n`);
+    for (const [folderName, icon] of Object.entries(icons.light.folderNames)) {
+        folderIcons.write(`\t'${folderName}': '${getIconPath(icon)}',\n`);
+    }
+    folderIcons.write("};\n");
+    folderIcons.end();
+})();
 
-    // FileNames to Icon
-    writeFile(`FileNamesToIcon`, () => {
-        const fileNames = Object.keys(icons.light.fileNames).reduce((acc, fileName) => ({
-            ...acc,
-            [fileName]: iconToPath[
-                icons.light.fileNames[fileName]
-            ]
-        }), {});
-        return fileNames;
-    });
-
-    // Languages to Icon
-    const vscodeLanguages = JSON.parse(vscodeLanguagesFile.toString()) as {
-        [languageId: string]: {
-            extensions: [string]
+// FileExtensions to Icon
+(async function() {
+    log.debug("creating table for filextensions");
+    const fileExtensions1 = createWriteStream(PATH_ICONSDATA + 'FileExtensions1ToIcon.ts', { flags: 'w'});
+    const fileExtensions2 = createWriteStream(PATH_ICONSDATA + 'FileExtensions2ToIcon.ts', { flags: 'w'});
+    fileExtensions1.write(`export const FileExtensions1ToIcon: { [key: string]: string } = {\n`);
+    fileExtensions2.write(`export const FileExtensions2ToIcon: { [key: string]: string } = {\n`);
+    for (const [extension, icon] of Object.entries(icons.light.fileExtensions)) {
+        if (extension.indexOf(".") === -1) {
+            fileExtensions1.write(`\t'${extension}': '${getIconPath(icon)}',\n`);
+        } else {
+            fileExtensions2.write(`\t'${extension}': '${getIconPath(icon)}',\n`)
         }
-    };
-    writeFile(`LanguagesToIcon`, () => {
-        const languagesIds = Object.keys(languages).reduce((acc, languageId) => {
-            const language = languages[languageId];
-            const defaultExtension = language.defaultExtension;
-            const iconFileName = icons.languageIds[languageId];
 
-            // sometimes, icon for language not exists, so skip it
-            if (iconFileName === undefined) {
-                return {
-                    ...acc
-                };
+    }
+    fileExtensions1.write("};\n");
+    fileExtensions2.write("};\n");
+    fileExtensions1.end();
+    fileExtensions2.end();
+})();
+
+// FileNames to Icon
+(async function() {
+    log.debug("creating table for filenames");
+    const alreadyIncludedNames: string[] = [];
+    const filenames = createWriteStream(PATH_ICONSDATA + 'FileNamesToIcon.ts', { flags: 'w' });
+    filenames.write(`export const FileNamesToIcon: { [key: string]: string } = {\n`);
+    // get folder names from vsi langauges definitions
+    for (const [filename, icon] of Object.entries(icons.light.fileNames)) {
+        filenames.write(`\t'${filename}': '${getIconPath(icon)}',\n`);
+        alreadyIncludedNames.push(filename);
+    }
+    // get folder names from vscode languages definitions
+    for (const [language, data] of Object.entries(vscodeLanguages)) {
+        if (data.filenames && Object.keys(vsiLanguages).includes(language) && getIconPath(`_f_${language}`)) {
+            for(const filename of data.filenames) {
+                if (!alreadyIncludedNames.includes(filename)) {
+                    filenames.write(`\t'${filename}': '${getIconPath(`_f_${language}`)}',\n`);
+                    alreadyIncludedNames.push(filename);
+                }
             }
+        }
+    }
+    filenames.write("};\n");
+    filenames.end();
+})();
 
-            // light theme version should be
-            const withoutPrefix = iconFileName.slice(3); // remove prefix "_f_"
-            const lightIconFileName = `_f_light_${withoutPrefix}`;
-            const existsLightTheme = icons.iconDefinitions[lightIconFileName]; // try to find light theme of icon
+// Languages to Icon
+(async function() {
+    log.debug("creating table for languages");
+    const alreadyIncludedLangs: string[] = [];
+    const languages = createWriteStream(PATH_ICONSDATA + 'LanguagesToIcon.ts', { flags: 'w' });
+    languages.write(`export const LanguagesToIcon: { [key: string]: string } = {\n`);
+    for (const [language, icon] of Object.entries(vsiLanguages)) {
+        const iconFileName = icons.languageIds[language]
+        if (iconFileName) {
+            const withoutPrefix = iconFileName.slice(3); // remove prefix "_f_";
+            const lightIconFilename = `_f_light_${withoutPrefix}`;
+            const existsLightTheme = icons.iconDefinitions[lightIconFilename]; // try to find light theme of icon
             const iconPath = existsLightTheme
-                ? iconToPath[lightIconFileName]
-                : iconToPath[iconFileName];
+                ? getIconPath(lightIconFilename)
+                : getIconPath(iconFileName);
 
             // Are there any language extensions supported by vscode ?
-            if (vscodeLanguages[languageId]) {
-                const supportedExtensions = vscodeLanguages[languageId].extensions;
+            if (vscodeLanguages[language]) {
+                const supportedExtensions = vscodeLanguages[language].extensions;
                 const languageExtensions: { [ext: string]: string } = {};
-                supportedExtensions.forEach((extension) => {
-                    // slice(1) - remove dot (e.g. ".cpp" to "cpp")
-                    languageExtensions[extension.slice(1)] = iconPath;
-                });
+                for (const extension of supportedExtensions) {
+                    languageExtensions[extension.slice(1)] = iconPath; // .cpp => cpp
+                };
                 // Override default extension
-                languageExtensions[language.defaultExtension] = iconPath;
-                return {
-                    ...acc,
-                    ...languageExtensions
+                // languageExtensions[icon.defaultExtension] = iconPath;
+
+                for(const [extIcon, extIconPath] of Object.entries(languageExtensions)) {
+                    if (!alreadyIncludedLangs.includes(extIcon)) {
+                        languages.write(`\t'${extIcon}': '${extIconPath}',\n`);
+                        alreadyIncludedLangs.push(extIcon);
+                    }
                 }
             } else {
-                return {
-                    ...acc,
-                    [language.defaultExtension]: iconPath
-                };
+                if (!alreadyIncludedLangs.includes(icon.defaultExtension)) {
+                    languages.write(`\t'${icon.defaultExtension}': '${iconPath}',\n`);
+                    alreadyIncludedLangs.push(icon.defaultExtension);
+                }
             }
-
-        }, {});
-        return languagesIds;
-    });
-    exit();
-});
+        }
+        // languages.write(`\t"${filename}": "${getIconPath(icon)}",\n`);
+    }
+    languages.write("};\n");
+    languages.end();
+})();